Creating shared component for displaying number of hints used per strategy
